some more clang-tidy warnings

+++ src/dat/unit/Combat.cpp
@@ -65,7 +65,7 @@ void Combat::serializeObject()
             && (gv > GV_LatestTap || gv < GV_T3)) {
         int8_t defarmor_byte = BaseArmor;
         serialize<int8_t>(defarmor_byte);
-        BaseArmor = defarmor_byte;
+        BaseArmor = uint8_t(defarmor_byte);
     } else {
         serialize<int16_t>(BaseArmor);
     }
